refactor: clean unused scripts
