--- src/vaults/index.ts
@@ -1,1 +1,0 @@
-export { default as distributeVaults } from "./distributeVaults";
